requirements.txt
flet==0.22.*
peewee~=3.17.5
openpyxl==3.1.3
python-docx==1.1.2